fix(DataProtector): Don't compare b64Json identities (field order bug)
The _handleShareWithSelf code did not compare identities correctly.
Because identities are exposed as the base64 of a JSON string,
they cannot be directly compared (since the order of the fields
in the JSON object is nondeterministic).

This caused the user's identity to be included twice if mentionned
excplicitely in the recipients list, resulting in internal errors.