refactor(core): new local device keys creation scheme
The local device keys used to be created as session opening and persisted right away.

Now, we create them in memory at user or additional device creation, and persist them
locally, with the device id, only after the server response has been handled. Then, and
only then, can the local device be used as a completely initialized device. If the
process is interrupted before completion, the device keys will be overridden and a new
device will be created.