chore(core): make API calls cancelable during authentication
If a call to .close() occurs while the challenge is being retrieved,
e.g. if .close() follows a .start() call closely, the internal call
to retrieve the challenge will be aborted with OperationCanceled, and
we won't try to sign with an empty device signature key anymore.

Fixes the "Received INVALID or FORGED signature" errors triggered
server side.