chore(core): add preventive API call cancelation
Although we synchronously remove the reference to the session when
calling tanker.stop(), internal asynchronous processes might still
be in progress and lead to new API calls (e.g. a database call ends,
the process resumes and calls an API).

So let's be extra safe and short-circuit new API calls as soon as
the current session starts its closing process.