refactor(core): handle unrecoverable errors
In some cases, the server is able to notify us that the local information
is corrupted and should be wiped. This patch implements this behavior and
also stops the session when trying to create a device (registerIdentity()
or verifyIdentity() calls) and an error not related to the verification
itself occurs.