feat(test-utils): ensure the silenced methods are always restored
By using the wrapper() method, the silenced methods are always restored
at the end of the execution of the (async) method wrapped.

This avoids "Attempted to wrap <method_name> which is already wrapped"
errors thrown by sinon if trying to rewrap a non-restored method (in
case a previously failing test does not restore its own wrappings).