fix(bench): Nudge yarn to not swallow karma's end-of-options flag
When Yarn sees a --, it thinks everything before the -- is for yarn.
So `yarn exec gdb --args curl` passes --args to gdb,
but `yarn exec gdb --args curl -- --foo` passes --args to yarn,
and passes --foo to gdb.

BUT, if we do `yarn exec -- gdb --args curl -- --foo`, now yarn
takes everything before the first `--`, which is nothing,
and forwards the other options to the actual program we care about.