feat: Trigger bridge jobs directly on matching branch
Previously, we would detect the upstream branch and try to switch
the branch in the downstream in a run-ci script.
Doing it in the upstream is cleaner, as all the variables for the
downstream pipeline will be actually set to the right branch.

For native bindings, this also fixes mistagged releases