feat: Add Sentry tags with Tanker info on outbound exceptions
We now take an (optional) Sentry Hub object as Tanker option.
When an exception bubbles up through a running Session,
we add tags to the current Sentry Hub before re-throwing it.

If the option is left undefined, we try to get Sentry from globalThis
(but it can be explicitly set to null to disable our Sentry usage)