chore(ci): add @tanker/datastore-dexie-memory to the list of published packages
Yes, this is basically an internal package that we use for our tests, but it's also useful to our user for a) their tests and b) their runtime sometimes when they don't need caching at all. No guarantees on that last one, but we don't actually lose anything by publishing this package.