feat: Return explicit NetworkError when intercepted by a middlebox
We can't do things like look at the certificate in JS to know that the
request was intercepted, but if we detect a non-JSON response,
we know that it must come from an intermediary on the path
(proxy, middlebox, cloud gateway, ...)

In particular some prats are stuck behind ZScaler, and it sometimes
block Tanker decrypt calls. Today this is accounted as a Tanker internal
error and creates a Sentry alert for us.
The new NetworkError message should allow support to identify the issue,
and potentially help the client identity their network as the problem.